+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Release History
+* **0.2.0 - auto-generate popups based on triggers - 10/30/2025**
+    * 💡 `FEATURE` Added an option to initialize as tooltip or dialog
+    * 💡 `FEATURE` Added an `init` function which initializes a popup for each trigger
 * **0.1.1 - 10/29/2025**
     * 🌀 `ENHANCEMENT` Adds errors when something goes wrong
 * **0.1.0 - 10/28/2025**